+++ torchbit/__init__.py
@@ -2,3 +2,4 @@ from . import core
 from . import debug
 from . import tools
 from . import utils
+from . import runner