+++ torchbit/tools/__init__.py
@@ -3,3 +3,5 @@ from .shape_process import *
 from .transform import *
 from .port import *
 from .bit_struct import *
+from .sender_collector import *
+from .sender_collector import *